+++ bbpress.php
@@ -1,7 +1,12 @@
 <?php
 /* BBPress Template */
 
-get_header();
+// Show Header?
+if (esc_html( get_theme_mod('header_ap') == "1") ) {
+	get_header('title');
+} else {
+	get_header();
+}
 
 // determine main column size from active sidebars
 //$main_column_size = bootstrapBasicGetMainColumnSize();