+++ image.php
@@ -5,7 +5,13 @@
  * @package bootstrap-basic
  */
 
-get_header();
+// Show Header?
+if (esc_html( get_theme_mod('header_ap') == "1") ) {
+	get_header('title');
+} else {
+	get_header();
+}
+
 ?> 
 				<div class="col-md-12 content-area image-attachment" id="main-column">
 					<main id="main" class="site-main" role="main">